# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TurbineOps REST API
  description: Turbine inspection and maintenance management API
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:4000/api
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Bearer token obtained from /auth/login endpoint'
  schemas:
...
    Turbine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        manufacturer:
          type: string
        mwRating:
          type: number
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /turbines:
    get:
      summary: List turbines
      tags: [Turbines]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of results to skip
      responses:
        '200':
          description: Turbines list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Turbine'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized

    post:
      summary: Create turbine
      tags: [Turbines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "Turbine-001"
                manufacturer:
                  type: string
                  example: "GE Wind"
                mwRating:
                  type: number
                  description: Power rating in megawatts
                  example: 2.5
                lat:
                  type: number
                  description: Latitude coordinate
                  example: 40.7128
                lng:
                  type: number
                  description: Longitude coordinate
                  example: -74.0060
      responses:
        '201':
          description: Turbine created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turbine'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role
  /turbines/{id}:
    get:
      summary: Get turbine details
      tags: [Turbines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Turbine details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turbine'
        '401':
          description: Unauthorized
        '404':
          description: Turbine not found

    put:
      summary: Update turbine
      tags: [Turbines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                manufacturer:
                  type: string
                mwRating:
                  type: number
                lat:
                  type: number
                lng:
                  type: number
      responses:
        '200':
          description: Turbine updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role
        '404':
          description: Turbine not found

    delete:
      summary: Delete turbine
      tags: [Turbines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Turbine deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN role
        '404':
          description: Turbine not found
...
  /inspections/{inspectionId}/repair-plan:
    post:
      summary: Generate repair plan
      tags: [RepairPlan]
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Repair plan generated
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN or ENGINEER role

    get:
      summary: Get repair plan
      tags: [RepairPlan]
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Repair plan details
        '401':
          description: Unauthorized
        '404':
          description: Repair plan not found

    put:
      summary: Update repair plan
      tags: [RepairPlan]
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priority:
                  type: string
                  enum: [LOW, MEDIUM, HIGH]
                estimatedCost:
                  type: number
      responses:
        '200':
          description: Repair plan updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN role

    delete:
      summary: Delete repair plan
      tags: [RepairPlan]
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Repair plan deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires ADMIN role